[DEV-1967] rename enum file
