+++ src/subdomains/core/buy-crypto/routes/buy/dto/get-buy-payment-info.dto.ts
@@ -17,7 +17,7 @@ import { Fiat } from 'src/shared/models/fiat/fiat.entity';
 import { Util } from 'src/shared/utils/util';
 import { XOR } from 'src/shared/validators/xor.validator';
 import { IsDfxIban } from 'src/subdomains/supporting/bank/bank-account/is-dfx-iban.validator';
-import { FiatPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.dto';
+import { FiatPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.enum';
 
 export class GetBuyPaymentInfoDto {
   @ApiPropertyOptional()
