+++ src/subdomains/core/buy-crypto/routes/crypto-route/crypto-route.controller.ts
@@ -16,7 +16,7 @@ import { HistoryDto } from 'src/subdomains/core/history/dto/history.dto';
 import { FeeDirectionType } from 'src/subdomains/generic/user/models/user/user.entity';
 import { UserService } from 'src/subdomains/generic/user/models/user/user.service';
 import { DepositDtoMapper } from 'src/subdomains/supporting/address-pool/deposit/dto/deposit-dto.mapper';
-import { CryptoPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.dto';
+import { CryptoPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.enum';
 import { TransactionHelper } from 'src/subdomains/supporting/payment/services/transaction-helper';
 import { CryptoRoute } from './crypto-route.entity';
 import { CryptoRouteService } from './crypto-route.service';
