+++ src/subdomains/core/sell-crypto/route/sell.controller.ts
@@ -14,7 +14,7 @@ import { PaymentInfoService } from 'src/shared/services/payment-info.service';
 import { Util } from 'src/shared/utils/util';
 import { UserService } from 'src/subdomains/generic/user/models/user/user.service';
 import { DepositDtoMapper } from 'src/subdomains/supporting/address-pool/deposit/dto/deposit-dto.mapper';
-import { CryptoPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.dto';
+import { CryptoPaymentMethod } from 'src/subdomains/supporting/payment/dto/payment-method.enum';
 import { TransactionHelper } from 'src/subdomains/supporting/payment/services/transaction-helper';
 import { BuyFiatService } from '../process/buy-fiat.service';
 import { CreateSellDto } from './dto/create-sell.dto';
